+++ test/unit/textbox.js
@@ -67,7 +67,7 @@
       }
     ],
     minWidth: 20,
-    splitByGrapheme: false,
+    textOverflow: '',
     strokeUniform: false,
     path: null,
     direction: 'ltr',
@@ -270,28 +270,73 @@
     textbox.initDimensions();
     assert.equal(textbox.textLines[0], 'xa', 'first line match expectations spacing 800');
   });
-  QUnit.test('wrapping with charspacing and splitByGrapheme positive', function(assert) {
+  QUnit.test('wrapping with charspacing and textOverflow `anywhere` positive', function(assert) {
     var textbox = new fabric.Textbox('xaxbxcxdeyaybid', {
       width: 190,
-      splitByGrapheme: true,
+      textOverflow: 'anywhere',
       charSpacing: 400
     });
     assert.deepEqual(
       textbox.textLines,
       ['xaxbx', 'cxdey', 'aybid'],
-      'lines match splitByGrapheme charSpacing 400'
+      'lines match textOverflow `anywhere` charSpacing 400'
     );
   });
-  QUnit.test('wrapping with charspacing and splitByGrapheme negative', function(assert) {
+  QUnit.test('wrapping with charspacing and textOverflow `anywhere` negative', function(assert) {
     var textbox = new fabric.Textbox('xaxbxcxdeyaybid', {
       width: 190,
-      splitByGrapheme: true,
+      textOverflow: 'anywhere',
       charSpacing: -100
     });
     assert.deepEqual(
       textbox.textLines,
       ['xaxbxcxdeyay', 'bid'],
-      'lines match splitByGrapheme charSpacing -100'
+      'lines match textOverflow `anywhere` charSpacing -100'
+    );
+  });
+  QUnit.test('wrapping with charspacing and textOverflow `break-word` positive', function(assert) {
+    var textbox = new fabric.Textbox('xaxbxcxdeyaybid', {
+      width: 190,
+      textOverflow: 'break-word',
+      charSpacing: 400
+    });
+    assert.deepEqual(
+      textbox.textLines,
+      ['xaxbx', 'cxdey', 'aybid'],
+      'lines match textOverflow `break-word` charSpacing 400'
+    );
+  });
+  QUnit.test('wrapping with textOverflow `break-word` and whitespace', function (assert) {
+    var textbox = new fabric.Textbox(
+      'xaxbxc                       xdeyaybid                                                                     \n   ',
+      {
+        width: 190,
+        textOverflow: 'break-word',
+        charSpacing: 400
+      }
+    );
+    assert.deepEqual(
+      textbox.textLines,
+      [
+        'xaxbx',
+        'c                       ',
+        'xdeya',
+        'ybid                                                                     ',
+        '   '
+      ],
+      'lines match textOverflow `break-word'
+    );
+  });
+  QUnit.test('wrapping with charspacing and textOverflow `break-word` negative', function(assert) {
+    var textbox = new fabric.Textbox('xaxbxcxdeyaybid', {
+      width: 190,
+      textOverflow: 'break-word',
+      charSpacing: -100
+    });
+    assert.deepEqual(
+      textbox.textLines,
+      ['xaxbxcxdeyay', 'bid'],
+      'lines match textOverflow `break-word` charSpacing -100'
     );
   });
   QUnit.test('wrapping with different things', function(assert) {
@@ -306,24 +351,36 @@
     assert.equal(textbox.textLines[5], 'ya', '5 line match expectations');
     assert.equal(textbox.textLines[6], 'yb', '6 line match expectations');
   });
-  QUnit.test('wrapping with splitByGrapheme', function(assert) {
+  QUnit.test('wrapping with textOverflow `anywhere`', function(assert) {
     var textbox = new fabric.Textbox('xaxbxcxdxeyaybid', {
       width: 1,
-      splitByGrapheme: true,
+      textOverflow: 'anywhere',
     });
-    assert.equal(textbox.textLines[0], 'x', '0 line match expectations splitByGrapheme');
-    assert.equal(textbox.textLines[1], 'a', '1 line match expectations splitByGrapheme');
-    assert.equal(textbox.textLines[2], 'x', '2 line match expectations splitByGrapheme');
-    assert.equal(textbox.textLines[3], 'b', '3 line match expectations splitByGrapheme');
-    assert.equal(textbox.textLines[4], 'x', '4 line match expectations splitByGrapheme');
-    assert.equal(textbox.textLines[5], 'c', '5 line match expectations splitByGrapheme');
+    assert.equal(textbox.textLines[0], 'x', '0 line match expectations');
+    assert.equal(textbox.textLines[1], 'a', '1 line match expectations');
+    assert.equal(textbox.textLines[2], 'x', '2 line match expectations');
+    assert.equal(textbox.textLines[3], 'b', '3 line match expectations');
+    assert.equal(textbox.textLines[4], 'x', '4 line match expectations');
+    assert.equal(textbox.textLines[5], 'c', '5 line match expectations');
+  });
+  QUnit.test('wrapping with textOverflow `break-word`', function(assert) {
+    var textbox = new fabric.Textbox('xaxbxcxdxeyaybid', {
+      width: 1,
+      textOverflow: 'break-word',
+    });
+    assert.equal(textbox.textLines[0], 'x', '0 line match expectations');
+    assert.equal(textbox.textLines[1], 'a', '1 line match expectations');
+    assert.equal(textbox.textLines[2], 'x', '2 line match expectations');
+    assert.equal(textbox.textLines[3], 'b', '3 line match expectations');
+    assert.equal(textbox.textLines[4], 'x', '4 line match expectations');
+    assert.equal(textbox.textLines[5], 'c', '5 line match expectations');
   });
   QUnit.test('wrapping with custom space', function(assert) {
     var textbox = new fabric.Textbox('xa xb xc xd xe ya yb id', {
       width: 2000,
     });
     var line1 = textbox._wrapLine('xa xb xc xd xe ya yb id', 0, 100, 0);
-    var expected1 =  [
+    var expected1 = [
       ['x', 'a', ' ', 'x', 'b'],
       ['x', 'c', ' ', 'x', 'd'],
       ['x', 'e', ' ', 'y', 'a'],
@@ -348,10 +405,13 @@
       width: 10,
     });
     var line1 = textbox._wrapLine('', 0, 100, 0);
-    assert.deepEqual(line1, [[]], 'wrapping without splitByGrapheme');
-    textbox.splitByGrapheme = true;
+    assert.deepEqual(line1, [[]], 'wrapping without textOverflow `anywhere`');
+    textbox.textOverflow = 'anywhere';
     var line2 = textbox._wrapLine('', 0, 100, 0);
-    assert.deepEqual(line2, [[]], 'wrapping with splitByGrapheme');
+    assert.deepEqual(line2, [[]], 'wrapping with textOverflow `anywhere`');
+    textbox.textOverflow = 'break-word';
+    var line3 = textbox._wrapLineWithTextOverflow('', 0, 100, 0);
+    assert.deepEqual(line3, [[]], 'wrapping with textOverflow `break-word`');
   });
   QUnit.test('texbox will change width from the mr corner', function(assert) {
     var text = new fabric.Textbox('xa xb xc xd xe ya yb id', { strokeWidth: 0 });
@@ -418,9 +478,9 @@
     assert.equal(iText.styles[4], undefined, 'style line 4 has been removed');
   });
 
-  QUnit.test('get2DCursorLocation with splitByGrapheme', function(assert) {
+  QUnit.test('get2DCursorLocation with textOverflow `anywhere`', function(assert) {
     var iText = new fabric.Textbox('aaaaaaaaaaaaaaaaaaaaaaaa',
-      { width: 60, splitByGrapheme: true });
+      { width: 60, textOverflow: 'anywhere' });
     var loc = iText.get2DCursorLocation();
 
     // [ [ '由', '石', '墨' ],
@@ -456,9 +516,47 @@
     assert.equal(loc.charIndex, 2, 'selection end 14 char 2');
   });
 
-  QUnit.test('missingNewlineOffset with splitByGrapheme', function(assert) {
+  QUnit.test('get2DCursorLocation with textOverflow `break-word`', function(assert) {
+    var iText = new fabric.Textbox('aaaaaaaaaaaaaaaaaaaaaaaa',
+      { width: 60, textOverflow: 'break-word' });
+    var loc = iText.get2DCursorLocation();
+
+    // [ [ '由', '石', '墨' ],
+    //   [ '分', '裂', '的' ],
+    //   [ '石', '墨', '分' ],
+    //   [ '裂', '由', '石' ],
+    //   [ '墨', '分', '裂' ],
+    //   [ '由', '石', '墨' ],
+    //   [ '分', '裂', '的' ],
+    //   [ '石', '墨', '分' ],
+    //   [ '裂' ] ]
+
+    assert.equal(loc.lineIndex, 0);
+    assert.equal(loc.charIndex, 0);
+
+    // '由石墨|分裂的石墨分裂由石墨分裂由石墨分裂的石墨分裂'
+    iText.selectionStart = iText.selectionEnd = 4;
+    loc = iText.get2DCursorLocation();
+
+    assert.equal(loc.lineIndex, 1, 'selection end 4 line 1');
+    assert.equal(loc.charIndex, 1, 'selection end 4 char 1');
+
+    iText.selectionStart = iText.selectionEnd = 7;
+    loc = iText.get2DCursorLocation();
+
+    assert.equal(loc.lineIndex, 2, 'selection end 7 line 2');
+    assert.equal(loc.charIndex, 1, 'selection end 7 char 1');
+
+    iText.selectionStart = iText.selectionEnd = 14;
+    loc = iText.get2DCursorLocation();
+
+    assert.equal(loc.lineIndex, 4, 'selection end 14 line 4');
+    assert.equal(loc.charIndex, 2, 'selection end 14 char 2');
+  });
+
+  QUnit.test('missingNewlineOffset with textOverflow `anywhere`', function(assert) {
     var textbox = new fabric.Textbox('aaa\naaaaaa\na\naaaaaaaaaaaa\naaa',
-      { width: 80, splitByGrapheme: true });
+      { width: 80, textOverflow: 'anywhere' });
 
     // [ [ 'a', 'a', 'a' ],
     //   [ 'a', 'a', 'a', 'a' ],
@@ -486,6 +584,27 @@
     assert.equal(offset, 1, 'it returns always 1');
     var offset = texbox.missingNewlineOffset(2);
     assert.equal(offset, 1, 'it returns always 1');
+  });
+
+  QUnit.test('missingNewlineOffset with textOverflow `break-word`', function(assert) {
+    var textbox = new fabric.Textbox('aaa\naaaaaa\na\naaaaaaaaaaaa\naaa',
+      { width: 80, textOverflow: 'break-word' });
+
+    // Same behavior as textOverflow `anywhere` in this case
+    // [ [ 'a', 'a', 'a' ],
+    //   [ 'a', 'a', 'a', 'a' ],
+    //   [ 'a', 'a' ],
+    //   [ 'a' ],
+    //   [ 'a', 'a', 'a', 'a' ],
+    //   [ 'a', 'a', 'a', 'a' ],
+    //   [ 'a', 'a', 'a', 'a' ],
+    //   [ 'a', 'a', 'a' ] ]
+
+    var offset = textbox.missingNewlineOffset(0);
+    assert.equal(offset, 1, 'line 0 is interrupted by a \n so has an offset of 1');
+
+    offset = textbox.missingNewlineOffset(1);
+    assert.equal(offset, 0, 'line 1 is wrapped without a \n so it does have an extra char count');
   });
 
   QUnit.test('_getLineStyle', function(assert) {
@@ -530,7 +649,6 @@
     var styles = {};
     for (var index = 0; index < text.length; index++) {
       styles[index] = { fontSize: 4 };
-      
     }
     var textbox = new fabric.Textbox(text, {
       styles: { 0: styles },
@@ -584,12 +702,25 @@
     var str = '0123456789';
     var measureTextbox = new fabric.Textbox(str, {
       fontSize: 20,
-      splitByGrapheme: false,
+      textOverflow: '',
     });
     var newTextbox = new fabric.Textbox(str, {
       width: measureTextbox.width,
       fontSize: 20,
-      splitByGrapheme: true,
+      textOverflow: 'anywhere',
+    });
+    assert.equal(newTextbox.textLines.length, measureTextbox.textLines.length, 'The same text is not wrapped');
+  });
+  QUnit.test('The same text does not need to be wrapped.', function(assert) {
+    var str = '0123456789';
+    var measureTextbox = new fabric.Textbox(str, {
+      fontSize: 20,
+      textOverflow: '',
+    });
+    var newTextbox = new fabric.Textbox(str, {
+      width: measureTextbox.width,
+      fontSize: 20,
+      textOverflow: 'break-word',
     });
     assert.equal(newTextbox.textLines.length, measureTextbox.textLines.length, 'The same text is not wrapped');
   });